updates to support taxes

+++ erp_products/db/migrate/20150507075057_add_dimensions_to_product_types.rb
@@ -1,10 +1,12 @@
 class AddDimensionsToProductTypes < ActiveRecord::Migration
   def change
-  	add_column :product_types, :length, :decimal
-  	add_column :product_types, :width, :decimal
-  	add_column :product_types, :height, :decimal
-  	add_column :product_types, :weight, :decimal
-  	add_column :product_types, :cylindrical, :boolean
-  	remove_column :product_types, :shipping_cost
+    unless column_exists? :product_types, :length
+      add_column :product_types, :length, :decimal
+      add_column :product_types, :width, :decimal
+      add_column :product_types, :height, :decimal
+      add_column :product_types, :weight, :decimal
+      add_column :product_types, :cylindrical, :boolean
+      remove_column :product_types, :shipping_cost
+    end
   end
 end
